Separated scraper from storage and tweaked comment format
